Move /lib files to /src directory
