+++ src/Dialog.hpp
